@@ -133,9 +133,7 @@ namespace Microsoft.Data.SqlClient.ConnectionPool
                     if (pool != null)
                     {
                         DbConnectionFactory connectionFactory = pool.ConnectionFactory;
-#if NETFRAMEWORK
-                        connectionFactory.PerformanceCounters.NumberOfActiveConnectionPools.Decrement();
-#endif
+
                         connectionFactory.QueuePoolForRelease(pool, true);
                     }
                 }
@@ -206,10 +204,8 @@ namespace Microsoft.Data.SqlClient.ConnectionPool
                                     newPool.Startup(); // must start pool before usage
                                     bool addResult = _poolCollection.TryAdd(currentIdentity, newPool);
                                     Debug.Assert(addResult, "No other pool with current identity should exist at this point");
-                                    SqlClientEventSource.Log.EnterActiveConnectionPool();
-#if NETFRAMEWORK
-                                    connectionFactory.PerformanceCounters.NumberOfActiveConnectionPools.Increment();
-#endif
+                                    SqlClientEventSource.Metrics.EnterActiveConnectionPool();
+
                                     pool = newPool;
                                 }
                                 else
@@ -285,9 +281,7 @@ namespace Microsoft.Data.SqlClient.ConnectionPool
                                 // pool into a list of pools to be released when they
                                 // are completely empty.
                                 DbConnectionFactory connectionFactory = pool.ConnectionFactory;
-#if NETFRAMEWORK
-                                connectionFactory.PerformanceCounters.NumberOfActiveConnectionPools.Decrement();
-#endif
+
                                 connectionFactory.QueuePoolForRelease(pool, false);
                             }
                             else
