@@ -13,6 +13,7 @@ using System.Runtime.Versioning;
 using System.Security;
 using System.Security.Permissions;
 using Microsoft.Data.Common;
+using Microsoft.Data.Common.ConnectionString;
 using Microsoft.Data.SqlClient.LocalDb;
 
 namespace Microsoft.Data.SqlClient
@@ -23,6 +24,7 @@ namespace Microsoft.Data.SqlClient
         // used by pooling classes so it is much easier to verify correctness
         // when not worried about the class being modified during execution
 
+        // @TODO: Remove this in favor of using DbConnectionStringDefaults??
         internal static class DEFAULT
         {
             internal const ApplicationIntent ApplicationIntent = DbConnectionStringDefaults.ApplicationIntent;
@@ -71,6 +73,7 @@ namespace Microsoft.Data.SqlClient
 #endif // NETFRAMEWORK
         }
 
+        // @TODO: Remove in favor of DbConnectionStringKeywords
         // SqlConnection ConnectionString Options
         internal static class KEY
         {
@@ -126,6 +129,7 @@ namespace Microsoft.Data.SqlClient
 #endif // NETFRAMEWORK
         }
 
+        // @TODO: Remove in favor DbConnectionStringSynonyms
         // Constant for the number of duplicate options in the connection string
         private static class SYNONYM
         {
@@ -575,22 +579,22 @@ namespace Microsoft.Data.SqlClient
 
             if (Authentication == SqlAuthenticationMethod.ActiveDirectoryManagedIdentity && _hasPasswordKeyword)
             {
-                throw SQL.NonInteractiveWithPassword(DbConnectionStringBuilderUtil.ActiveDirectoryManagedIdentityString);
+                throw SQL.NonInteractiveWithPassword(DbConnectionStringUtilities.ActiveDirectoryManagedIdentityString);
             }
 
             if (Authentication == SqlAuthenticationMethod.ActiveDirectoryMSI && _hasPasswordKeyword)
             {
-                throw SQL.NonInteractiveWithPassword(DbConnectionStringBuilderUtil.ActiveDirectoryMSIString);
+                throw SQL.NonInteractiveWithPassword(DbConnectionStringUtilities.ActiveDirectoryMSIString);
             }
 
             if (Authentication == SqlAuthenticationMethod.ActiveDirectoryDefault && _hasPasswordKeyword)
             {
-                throw SQL.NonInteractiveWithPassword(DbConnectionStringBuilderUtil.ActiveDirectoryDefaultString);
+                throw SQL.NonInteractiveWithPassword(DbConnectionStringUtilities.ActiveDirectoryDefaultString);
             }
 
             if (Authentication == SqlAuthenticationMethod.ActiveDirectoryWorkloadIdentity && _hasPasswordKeyword)
             {
-                throw SQL.NonInteractiveWithPassword(DbConnectionStringBuilderUtil.ActiveDirectoryWorkloadIdentityString);
+                throw SQL.NonInteractiveWithPassword(DbConnectionStringUtilities.ActiveDirectoryWorkloadIdentityString);
             }
         }
 
@@ -941,7 +945,7 @@ namespace Microsoft.Data.SqlClient
             // wrap Format and Overflow exceptions with Argument one, to be consistent with rest of the keyword types (like int and bool)
             try
             {
-                return DbConnectionStringBuilderUtil.ConvertToApplicationIntent(KEY.ApplicationIntent, value);
+                return DbConnectionStringUtilities.ConvertToApplicationIntent(KEY.ApplicationIntent, value);
             }
             catch (FormatException e)
             {
@@ -973,7 +977,7 @@ namespace Microsoft.Data.SqlClient
 
             try
             {
-                return DbConnectionStringBuilderUtil.ConvertToAuthenticationType(KEY.Authentication, value);
+                return DbConnectionStringUtilities.ConvertToAuthenticationType(KEY.Authentication, value);
             }
             catch (FormatException e)
             {
@@ -998,7 +1002,7 @@ namespace Microsoft.Data.SqlClient
 
             try
             {
-                return DbConnectionStringBuilderUtil.ConvertToColumnEncryptionSetting(KEY.ColumnEncryptionSetting, value);
+                return DbConnectionStringUtilities.ConvertToColumnEncryptionSetting(KEY.ColumnEncryptionSetting, value);
             }
             catch (FormatException e)
             {
@@ -1023,7 +1027,7 @@ namespace Microsoft.Data.SqlClient
 
             try
             {
-                return DbConnectionStringBuilderUtil.ConvertToAttestationProtocol(KEY.AttestationProtocol, value);
+                return AttestationProtocolUtilities.ConvertToAttestationProtocol(KEY.AttestationProtocol, value);
             }
             catch (FormatException e)
             {
@@ -1048,7 +1052,7 @@ namespace Microsoft.Data.SqlClient
 
             try
             {
-                return DbConnectionStringBuilderUtil.ConvertToIPAddressPreference(KEY.IPAddressPreference, value);
+                return IpAddressPreferenceUtilities.ConvertToIPAddressPreference(KEY.IPAddressPreference, value);
             }
             catch (FormatException e)
             {
@@ -1069,7 +1073,7 @@ namespace Microsoft.Data.SqlClient
 
             try
             {
-                return DbConnectionStringBuilderUtil.ConvertToPoolBlockingPeriod(KEY.PoolBlockingPeriod, value);
+                return PoolBlockingUtilities.ConvertToPoolBlockingPeriod(KEY.PoolBlockingPeriod, value);
             }
             catch (Exception e) when (e is FormatException || e is OverflowException)
             {
@@ -1086,7 +1090,7 @@ namespace Microsoft.Data.SqlClient
 
             try
             {
-                return DbConnectionStringBuilderUtil.ConvertToSqlConnectionEncryptOption(KEY.Encrypt, value);
+                return AttestationProtocolUtilities.ConvertToSqlConnectionEncryptOption(KEY.Encrypt, value);
             }
             catch (FormatException e)
             {
