@@ -4,13 +4,14 @@
 
 using System;
 using System.Diagnostics;
+using System.IO;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.Data.Common;
 
 namespace Microsoft.Data.SqlClient
 {
-    sealed internal class SqlSequentialStream : System.IO.Stream
+    internal sealed class SqlSequentialStream : Stream
     {
         private SqlDataReader _reader;  // The SqlDataReader that we are reading data from
         private readonly int _columnIndex;       // The index of out column in the table
@@ -302,7 +303,7 @@ namespace Microsoft.Data.SqlClient
             Task readTask = ReadAsync(buffer, offset, count, CancellationToken.None);
             if (callback != null)
             {
-                readTask.ContinueWith((t) => callback(t), TaskScheduler.Default);
+                readTask.ContinueWith(t => callback(t), TaskScheduler.Default);
             }
             return readTask;
         }
@@ -328,11 +329,19 @@ namespace Microsoft.Data.SqlClient
             return readTask.Result;
         }
 #else
-        public override IAsyncResult BeginRead(byte[] array, int offset, int count, AsyncCallback asyncCallback, object asyncState) =>
-            TaskToApm.Begin(ReadAsync(array, offset, count, CancellationToken.None), asyncCallback, asyncState);
+        public override IAsyncResult BeginRead(
+            byte[] array,
+            int offset,
+            int count,
+            AsyncCallback asyncCallback,
+            object asyncState)
+        {
+            Task<int> readTask = ReadAsync(array, offset, count, CancellationToken.None);
+            return TaskToAsyncResult.Begin(readTask, asyncCallback, asyncState);
+        }
 
         public override int EndRead(IAsyncResult asyncResult) =>
-            TaskToApm.End<int>(asyncResult);
+            TaskToAsyncResult.End<int>(asyncResult);
 #endif
     }
 }
