@@ -1,8 +1,14 @@
-﻿using System;
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+#if NETFRAMEWORK
+
+using System;
 using System.Buffers;
 using System.Text;
 
-namespace Microsoft.Data.SqlClient
+namespace Microsoft.Data.SqlClient.Utilities
 {
     internal static class BufferWriterExtensions
     {
@@ -24,3 +30,5 @@ namespace Microsoft.Data.SqlClient
         }
     }
 }
+
+#endif
