@@ -167,7 +167,7 @@ namespace Microsoft.Data.SqlClient
         public HealthReport(byte[] payload)
         {
             Size = payload.Length;
-#if NET9_0
+#if NET9_0_OR_GREATER
             Certificate = X509CertificateLoader.LoadCertificate(payload);
 #else
             Certificate = new X509Certificate2(payload);
