@@ -8326,7 +8326,7 @@ namespace System {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to .
+        ///   Looks up a localized string similar to Socket is null.
         /// </summary>
         internal static string SNI_ERROR_10 {
             get {
